프로젝트 모듈 단위로 분리
# 공통로직 구성 시 RDB, NOSQL 구성 시 동일한 VO객체를 사용 할 수 없어서 패키지 분리가 필요한 상황에 이유로 프로젝트 분리 작업
 - 공통로직 상송, 인터페이스 및 공통로직으로 사용 될 module-common프로젝트 분리
 - jpa용, mongo용의 프로젝트를 구성할 수 있도록 module-jpa프로젝트 분리 (mongo구현 예정)
 - web로직이 구성되어있는 module-web프로젝트 분리

# 모듈 구성에 따른 build.gradle 재 구성

# 페이징 처리 시 html에서 처리가 안되어있어서 controller임의 소스 반영
