+++ module-web/src/main/java/com/ymwoo/project/StandardApplication.java
