--- src/main/java/com/ymwoo/project/common/SearchRepository.java
@@ -1,36 +1,0 @@
-package com.ymwoo.project.common;
-
-import java.util.List;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.repository.NoRepositoryBean;
-
-@NoRepositoryBean
-public interface SearchRepository<Entity, Search extends AbstractSearch>
-{
-    List<Entity> searchList(Search search);
-
-
-
-
-
-    Entity searchOne(Search search);
-
-
-
-
-
-    Page<Entity> searchPage(Search search);
-
-
-
-
-
-    Long count(Search search);
-
-
-
-
-
-    Boolean exists(Search search);
-}
