--- src/test/java/com/ymwoo/project/StandardApplicationTests.java
@@ -1,15 +1,0 @@
-package com.ymwoo.project;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class StandardApplicationTests
-{
-
-    @Test
-    void contextLoads()
-    {
-    }
-
-}
